requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
pymysql
python-dotenv
python-jose
python-multipart